added multiple camera selection, preparation for shadow map impl

+++ opengl_renderer/include/Graphics/GPUBuffers.h
@@ -26,7 +26,7 @@ namespace Real {
     struct MaterialSSBO {
         glm::vec4 baseColor = glm::vec4(1.0, 0.2, 0.0, 1.0);
         glm::vec4 emissiveMetallic = glm::vec4(0.2f);
-        float textureLayers[4]; // 0 = tex diffuse layer, 1 = tex specular layer, other indices padding (16-byte alignment)
+        int textureLayers[4]; // 0 = tex diffuse layer, 1 = tex specular layer, other indices padding (16-byte alignment)
         float roughnessShininess[4]; // 0 = roughness, 1 = shininess
     };
